mqtt/requirements.txt
paho-mqtt
json
